+++ src/components/icon/svgs.tsx
@@ -2,6 +2,7 @@ import { ReactComponent as Backward } from "./svg/backward.svg";
 import { ReactComponent as Heart } from "./svg/heart.svg";
 import { ReactComponent as Home } from "./svg/home.svg";
 import { ReactComponent as Minus } from "./svg/minus.svg";
+import { ReactComponent as Palette } from "./svg/palette.svg";
 import { ReactComponent as Plus } from "./svg/plus.svg";
 import { ReactComponent as Star } from "./svg/star.svg";
 import { ReactComponent as Undo } from "./svg/undo.svg";
@@ -11,6 +12,7 @@ const svgs: any = {
   Heart,
   Home,
   Minus,
+  Palette,
   Plus,
   Star,
   Undo